Create Radio for Mainboard
some code that can get past init() for about 10% of the time. I think the problem is in the pin configuration... some pins are defaulted to certain modes. Also, We may need to dig into the default SPI.begin() function. It is being called by the RHHardwareSPI function spi.begin()... The rf.init() is failing at this case: // Get the device type and check it
    // This also tests whether we are really connected to a device
    _deviceType = spiRead(RH_RF22_REG_00_DEVICE_TYPE);
    if (   _deviceType != RH_RF22_DEVICE_TYPE_RX_TRX
        && _deviceType != RH_RF22_DEVICE_TYPE_TX)